refactor(web): new offers for sql order page
ref: MANAGER-11000
