+++ packages/manager/apps/web/client/app/hosting/database/order/public/components/steps/db-categories-offers/component.js
