+++ packages/manager/apps/web/client/app/hosting/database/order/public/components/steps/db-categories-offers/translations/Messages_fr_FR.json
